RDCC-252 refactored feign package name
